Bill Details Implementation
Users can enter a description of their bill, set a deadline and select who is being paid and who the payment for the bill is being sent from. Was able to fetch the proper information from the AddBillsPage through routing and display it properly in the dropdowns, etc. Saving and storing of these details are not able to be stored in the database yet since this screen is not finished completely, however the details entered in the previous screen AddBillsPage are stored properly in the database.