Amount Box Functionality Added
Created functionality so that the amounts entered by the selected participants must match the total bill amount. Also added the equal split function so that if that option is selected, no amounts are needed to be entered and the bill gets divided equally. Bill also submits and stores itself in database successfully still.